Eliminate GetOperators
Consolidate all logic in QueryOperators, which are now used for both `query` and `pull`.
